requirements.txt
requests
discord.py
aioconsole
ruamel.yaml
sre_yield
parse
psutil
